+++ SmolEngine/Source/Renderer/Vulkan/VulkanIndexBuffer.cpp
@@ -17,7 +17,7 @@ namespace SmolEngine
 
 	void VulkanIndexBuffer::Create(const uint32_t* data, uint64_t count)
 	{
-		m_ElementsCount = count;
+		m_ElementsCount = static_cast<uint32_t>(count);
 		m_IndexBuffer.Create(data, sizeof(uint32_t) * count, VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT | VK_MEMORY_PROPERTY_HOST_COHERENT_BIT, VK_BUFFER_USAGE_INDEX_BUFFER_BIT | VK_BUFFER_USAGE_TRANSFER_DST_BIT);
 	}
 
@@ -28,7 +28,7 @@ namespace SmolEngine
 
 	void VulkanIndexBuffer::SetData(const uint32_t* data, uint64_t count)
 	{
-		m_ElementsCount = count;
+		m_ElementsCount = static_cast<uint32_t>(count);
 		m_IndexBuffer.SetData(data, sizeof(uint32_t) * count);
 	}
